@@ -15,6 +15,7 @@ dependencies:
   #extended_image: ^1.3.0
   #flutter_mobx: ^1.1.0+2
   #get: ^3.26.0
+  #autocomplete_textfield: 1.7.3
   #metadata: 1.0.2
   #mobx: ^1.2.1+4
   #platform_alert_dialog: ^1.0.0+2
@@ -24,7 +25,6 @@ dependencies:
   #tflite_flutter: ^0.5.0
   #tflite: ^1.1.1
   after_layout: 1.1.0
-  #autocomplete_textfield: 1.7.3
   autocomplete_textfield:
     git:
       url: https://github.com/justkawal/flutter-autocomplete-textfield.git
@@ -37,12 +37,12 @@ dependencies:
   cloud_functions: 1.1.2
   collection: 1.15.0
   cryptography_flutter: 2.0.0
+  cryptography: 2.0.0
   cupertino_icons: 1.0.3
   date_utils: 0.2.0
   diacritic: 0.1.3
   email_validator: 2.0.1
   encrypt: 5.0.0
-  #esys_flutter_share: ^1.0.2
   expandable: 5.0.1
   extended_image: 4.1.0
   facebook_app_events: 0.12.0
@@ -57,13 +57,11 @@ dependencies:
   flutter_animator: 3.1.1
   flutter_blurhash: 0.6.0
   flutter_branch_sdk: 3.2.0
-  #flutter_device_locale: 0.4.0
   flutter_device_locale:
     git:
       url: https://github.com/justkawal/flutter_device_locale.git
   flutter_inapp_purchase: 5.0.1
   flutter_keyboard_visibility: 5.0.2
-  flutter_local_notifications: 7.0.0-dev.1
   flutter_secure_storage: 4.2.0
   flutter_spinkit: 5.0.0
   flutter_staggered_grid_view: 0.4.0
@@ -77,7 +75,7 @@ dependencies:
       url: https://github.com/justkawal/flutter_vibrate.git
   geocoding: 2.0.0
   geolocator: 7.1.0
-  get: 4.1.3
+  get: 4.1.4
   google_maps_flutter: 2.0.6
   hive: 2.0.4
   home_widget: 0.1.2+1
@@ -94,7 +92,7 @@ dependencies:
   photo_manager: 1.2.2
   photo_view: 0.11.1
   pull_to_refresh: 2.0.0
-  purchases_flutter: 3.2.2
+  purchases_flutter: 3.3.0
   rate_my_app: 1.1.0+1
   share_extend: 2.0.0
   share: 2.0.4
@@ -114,6 +112,8 @@ dependencies:
 
 dependency_overrides:
   intl: ^0.17.0
+  flutter_local_notifications: ^7.0.0-dev.1
+  flutter_local_notifications_platform_interface: 4.0.0
   path_provider: ^2.0.0
   convert: 3.0.0
   hive_flutter: ^1.0.0
