@@ -1,5 +1,5 @@
 import 'package:cryptography/cryptography.dart';
 
 class SecretKeyEntity {
-  SecretKey secretKey;
+  SecretKey? secretKey;
 }
